Build (almost) all binaries against openssl and aws-lc
this patch introduces a build of almost all binaries against both
openssl and aws-lc.

This is done by renaming libssl to libssl-openssl and libssl-aws-lc
and changing the needed field in the elf binaries. Same goes for
libcrypto.

This works almost correctly because after stipping the libs, some
binaries raise an error such as:

  curl-openssl: error while loading shared libraries: libngtcp2_crypto_ossl-openssl.so: ELF load command address/offset not page-aligned

The root cause still needs to be identified. For now a couple of libs
are not stripped to prevent from having this issue.

Some test still need to be run once all binaries are built.

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>